requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
easyocr==1.7.1
filelock==3.14.0
fsspec==2024.5.0
imageio==2.34.1
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pyclipper==1.3.0.post5
python-bidi==0.4.2
PyYAML==6.0.1
scikit-image==0.23.2
scipy==1.13.0
shapely==2.0.4
six==1.16.0
sqlparse==0.5.0
sympy==1.12
tifffile==2024.5.10
torch==2.3.0
torchvision==0.18.0
triton==2.3.0
typing_extensions==4.11.0
